created a game over sequence for bounds too fast on landing pad
revised topo polygon by setting first and last topo point as the corners of the screen
created a delayed level completion logic for landing on the pad

a continued issue is that the clearance area to read the acceptable speed of entry needs to be big enough to read the player's speed, but small enough to look like they have actually landed before analysing.
if it is too small and you coming in too fast in the y, I think the collision with the pad turns the y velocity to zero, and the clearance area does not have time to register the entry velocity.
I made the area larger and it works as expected now, but you can just hover over the pad without touching it and the level is completed.
next solution to look into: maybe analyze the speed of the lander while it is in the area, when it is 0,0, then level complete. However i dont want the velocity or rotation requirements to be needed in the area above the lander, only when they impact the landing site.